Expose different rotation strategies in REST API.
    Fixes #913.

+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-rest-client/src/main/java/org/graylog2/restclient/models/api/responses/system/indices/DeflectorConfigResponse.java
@@ -34,7 +34,7 @@ public class DeflectorConfigResponse {
     public Period maxTimePerIndex;
 
     @JsonProperty("max_size_per_index")
-    public long maxSizePerIndex;
+    public int maxSizePerIndex;
 
     @JsonProperty("max_number_of_indices")
     public int maxNumberOfIndices;
